update with weights = None and dtype float16

+++ pinns/pde_utils.py
@@ -1,7 +1,9 @@
 import torch
 
 
-def gen_uniform(bs, device, start=-1, end=1, temporal_scheme=False, M=None):
+def gen_uniform(
+    bs, device, start=-1, end=1, temporal_scheme=False, M=None, dtype=torch.float32
+):
     if temporal_scheme:
         step = (end - start) / M
         temporal_values = torch.arange(start, end + step, step=step)
@@ -12,7 +14,7 @@ def gen_uniform(bs, device, start=-1, end=1, temporal_scheme=False, M=None):
                 (r1 - r2)
                 * torch.rand(
                     (bs // M, 1),
-                    dtype=torch.float32,
+                    dtype=dtype,
                     device=device,
                     requires_grad=False,
                 )
@@ -21,6 +23,6 @@ def gen_uniform(bs, device, start=-1, end=1, temporal_scheme=False, M=None):
         vector = torch.cat(times, axis=0)
     else:
         vector = (end - start) * torch.rand(
-            (bs, 1), dtype=torch.float32, device=device, requires_grad=False
+            (bs, 1), dtype=dtype, device=device, requires_grad=False
         ) + start
     return vector
